@@ -19,23 +19,26 @@ version: 1.0.0+1
 
 environment:
 #  sdk: ">=2.7.0 <3.0.0"
-    sdk: ">=2.10.0 <3.0.0"
+#  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  url_launcher: ^5.4.0
-  fluttertoast: ^7.1.1
-  share: 0.6.5+2
-  firebase_core: "^0.5.0+1"
-  firebase_crashlytics: "^0.2.1+1"
-  firebase_database: "^4.1.1"
-  firebase_analytics: "^6.0.2"
-
-
+  url_launcher: ^6.0.20
+  fluttertoast: ^8.0.9
+  firebase_crashlytics: ^2.5.2
+  firebase_analytics: ^9.1.1
+  firebase_database: ^9.0.7
+  firebase_core: ^1.12.0
+  share: ^2.0.4
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.4
+
+dependency_overrides:
+  firebase_crashlytics_platform_interface: 3.1.13
+  firebase_analytics_platform_interface: 3.0.5
 
 dev_dependencies:
   flutter_test:
