@@ -13,7 +13,7 @@ class CurrencyConverterView extends StatefulWidget {
 }
 
 class _CurrencyConverterViewState extends State<CurrencyConverterView> {
-  Query _query;
+  Query? _query;
 
   @override
   void initState() {
@@ -35,61 +35,68 @@ class _CurrencyConverterViewState extends State<CurrencyConverterView> {
         padding: EdgeInsets.all(16.0),
         child: Column(
           children: [
-            StreamBuilder<DataSnapshot>(
-              stream: _query.once().asStream(),
-              builder: (context, snapshot) {
-                if (snapshot.hasError) {
-                  print('error in reading stream');
-                  return Container();
-                } else {
-                  Widget child;
-                  switch (snapshot.connectionState) {
-                    case ConnectionState.waiting:
-                      child = Container(child: Text('ConnectionState.waiting'));
-                      break;
-                    case ConnectionState.none:
-                      child = Container(child: Text('ConnectionState.none'));
-                      break;
-                    case ConnectionState.active:
-                      child = Container(child: Text('ConnectionState.active'));
-                      break;
-                    case ConnectionState.done:
-                      List<String> entries = List();
-                      snapshot.data.value
-                          .forEach((key, value) => {entries.add(key)});
-                      child = Container(
-                          color: Colors.transparent,
-                          child: Column(
-                            children: [
-                              Container(
-                                  child: TextBar(
-                                      type: "KB",
-                                      option: '1000',
-                                      leftText: 'From')),
-                              Container(
-                                margin: EdgeInsets.only(top: 40),
-                                child: DropDownView(
-                                    bgColor: Colors.grey.shade200,
-                                    entries: entries,
-                                    selected: entries[0],
-                                    onItemSelected: (value) {}),
-                              ),
-                              Container(
-                                margin: EdgeInsets.only(top: 12.0),
-                                child: ResultText("0000"),
-                              )
-                            ],
-                          ));
-                      break;
+            Builder(builder: (context) {
+              if (_query == null) {
+                return Container();
+              }
+              return StreamBuilder<DataSnapshot>(
+                // stream: _query!.once().asStream(),
+                builder: (context, snapshot) {
+                  if (snapshot.hasError) {
+                    print('error in reading stream');
+                    return Container();
+                  } else {
+                    Widget child;
+                    switch (snapshot.connectionState) {
+                      case ConnectionState.waiting:
+                        child =
+                            Container(child: Text('ConnectionState.waiting'));
+                        break;
+                      case ConnectionState.none:
+                        child = Container(child: Text('ConnectionState.none'));
+                        break;
+                      case ConnectionState.active:
+                        child =
+                            Container(child: Text('ConnectionState.active'));
+                        break;
+                      case ConnectionState.done:
+                        List<String> entries = [];
+                        // snapshot.data?.value
+                        //     .forEach((key, value) => {entries.add(key)});
+                        child = Container(
+                            color: Colors.transparent,
+                            child: Column(
+                              children: [
+                                Container(
+                                    child: TextBar(
+                                        type: "KB",
+                                        option: '1000',
+                                        leftText: 'From')),
+                                Container(
+                                  margin: EdgeInsets.only(top: 40),
+                                  child: DropDownView(
+                                      bgColor: Colors.grey.shade200,
+                                      entries: entries,
+                                      selected: entries[0],
+                                      onItemSelected: (value) {}),
+                                ),
+                                Container(
+                                  margin: EdgeInsets.only(top: 12.0),
+                                  child: ResultText("0000"),
+                                )
+                              ],
+                            ));
+                        break;
+                    }
+                    return Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [child],
+                    );
                   }
-                  return Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [child],
-                  );
-                }
-              },
-            )
+                },
+              );
+            })
           ],
         ),
       ),
