@@ -7,12 +7,17 @@ class ImageButton extends StatelessWidget {
   @required
   final onPress;
   EdgeInsets padding = EdgeInsets.all(0.0);
-  Color bgColor = Colors.red;
+  Color? bgColor = Colors.red;
 
   EdgeInsets defaultPadding = EdgeInsets.all(0.0);
   Color defaultBgColor = Colors.transparent;
 
-  ImageButton({this.path, this.onPress, this.padding, this.bgColor});
+  ImageButton({
+    required this.path,
+    this.onPress,
+    required this.padding,
+    this.bgColor,
+  });
 
   @override
   Widget build(BuildContext context) {
